adjust radius for poi

+++ app/src/main/java/com/github/mooncoders/jammed/sdk/PointsOfInterestParams.kt
@@ -3,5 +3,5 @@ package com.github.mooncoders.jammed.sdk
 data class PointsOfInterestParams(
     val latitude: Double,
     val longitude: Double,
-    val radius: Int
+    val radius: Double
 )
